refactor utils_wsi
